Add a test suite, ruff, and CI
160 tests over the two halves worth covering. The pure functions first, since
that is where data corruption is silent: specstruct's string <-> typed column
conversion, drivedb and ramdb, and entry's amount handling and expanders. The
drivedb cases are the notations the collection was actually recorded in, so a
change that mis-reads real data fails rather than quietly restructuring it,
and the RAM cases pin the parity arithmetic that makes eighteen 41256s 512 KB
rather than 576.

Then the behaviour that has actually broken this month: typed columns taking
blank form input, a select keeping a value outside its vocabulary, links
unlinking rather than dangling when their target is deleted, derived strings
never being written to directly, and a title that names the item.

The whole app runs against a throwaway SQLite database, built from the models
rather than from Alembic -- the migrations are MariaDB-specific on purpose
(REGEXP, STR_TO_DATE, MODIFY), and what is under test is the app.

Two things the tooling found immediately: datetime.utcnow() is deprecated and
due for removal, so the change log now uses an explicit naive UTC helper; and
ruff turned up twenty small things, of which the ones worth naming are a
missing `raise ... from` and a crop box using `l` for `left`.

ruff's config records why the noisy rules are off: the multiplication sign and
em dash are this catalogue's display typography, and FastAPI's Depends() is a
call in a default argument by design.

+++ api/app/entry.py
@@ -115,7 +115,7 @@ def to_kb(text: str):
     if unit not in _KB_UNITS:
         return None
     try:
-        return int(round(float(m.group(1)) * _KB_UNITS[unit]))
+        return round(float(m.group(1)) * _KB_UNITS[unit])
     except ValueError:
         return None
